+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.6
 
  - M8 Firmware 4.1 support
+ - Aligning with m8-files upstream
 
 # 0.5.1